chore: refactor util dir to testutil
